+++ src/components/Dashboard/DashboardHeader.tsx
@@ -1,0 +1,66 @@
+import React from 'react';
+import {View, Text, StyleSheet} from 'react-native';
+import MaterialCommunityIcons from 'react-native-vector-icons/MaterialCommunityIcons';
+import {formatDate} from '../../utils/date';
+import ProfileHeader from './header/ProfileHeader';
+import MobtiBox from './header/MobtiBox';
+import DriveInfoGroup from './header/DriveInfoBox';
+import {UseResponse} from '../../types/user';
+import {DashboardResponse} from '../../types/dashboard';
+
+export type DashboardHeaderProps = {
+  userInfo: UseResponse;
+  isEnabled: boolean;
+  toggleSwitch: () => void;
+  dashboard: DashboardResponse;
+};
+
+export default function DashboardHeader({
+  userInfo,
+  isEnabled,
+  toggleSwitch,
+  dashboard,
+}: DashboardHeaderProps) {
+  const driveInfoItems = [
+    {
+      icon: <MaterialCommunityIcons name="calendar-month" size={15} />,
+      label: '최근 운전일',
+      value: isNaN(new Date(dashboard?.lastDrive ?? '').getTime())
+        ? '-'
+        : formatDate(dashboard!.lastDrive),
+    },
+    {
+      icon: (
+        <MaterialCommunityIcons name="chart-bell-curve-cumulative" size={15} />
+      ),
+      label: '누적 운전 횟수',
+      value: dashboard?.driveCount ? `${dashboard.driveCount}회` : '-',
+    },
+  ];
+
+  return (
+    <>
+      <ProfileHeader
+        userInfo={userInfo}
+        isEnabled={isEnabled}
+        toggleSwitch={toggleSwitch}
+      />
+      <MobtiBox dashboard={dashboard} />
+      <DriveInfoGroup items={driveInfoItems} />
+      <View style={styles.reportBox}>
+        <Text style={styles.reportTitle}>운전 점수 리포트</Text>
+      </View>
+    </>
+  );
+}
+
+const styles = StyleSheet.create({
+  reportBox: {
+    marginTop: 40,
+    marginBottom: 10,
+  },
+  reportTitle: {
+    fontSize: 17,
+    fontWeight: 'bold',
+  },
+});
