+++ src/types/dashboard.ts
@@ -1,0 +1,31 @@
+import { UseResponse } from "./user";
+
+export type HomeStackParamList = {
+  Dashboard: undefined;
+  Feedback: undefined;
+};
+
+export type DashboardResponse = {
+  userId: string;
+  lastDrive: string;
+  driveCount: number;
+  scores: DriveScores;
+};
+
+export type DriveScores = {
+  idlingScore: number;
+  speedMaintainScore: number;
+  ecoScore: number;
+  accelerationScore: number;
+  sharpTurnScore: number;
+  overSpeedScore: number;
+  safetyScore: number;
+  reactionScore: number;
+  laneDepartureScore: number;
+  followingDistanceScore: number;
+  accidentPreventionScore: number;
+  drivingTimeScore: number;
+  inactivityScore: number;
+  attentionScore: number;
+  totalScore: number;
+};
