+++ src/utils/date.ts
@@ -1,0 +1,7 @@
+export const formatDate = (isoDate: string) => {
+  const date = new Date(isoDate);
+  const yyyy = date.getFullYear();
+  const mm = String(date.getMonth() + 1).padStart(2, '0');
+  const dd = String(date.getDate()).padStart(2, '0');
+  return `${yyyy}.${mm}.${dd}`;
+};
